feat: turn jinja extractor into a generic HTML template extractor
... supporting both jinja and JS microtemplates
